Add multiple files...
Add NFT.sol file to the contract folder and impoted the openzeppelin contract
the openzeppelin contracts enable the creation of contracts easily by simply importing them.
Add the deploy scripts to the deploy file to enable the contract deploy successufully.
Add the config file and populated it so that the contract will know where to deploy it.